Fixed a bug in authorizing a user to edit or delete posts.

+++ app/views/posts/show.blade.php
@@ -15,14 +15,15 @@
 				<h3 class="panel-title">{{{ $post->title }}}</h3>
 			</div>
 			<div class="panel-body">
-				{{{ $post->body }}}
+				<p>{{{ $post->body }}}</p>
+				<img src="{{{ $post->image }}}" alt="">
 			</div>
 			<div class="dateTime">
 				{{{ $post->created_at->setTimezone('America/Chicago')->format('l, F jS Y @ h:i:s A') }}}
 			</div>
 		</div>
 		<a href="{{{ action('PostsController@index') }}}"><span class="badge">Back to Index</span></a>
-		@if (Auth::check())
+		@if (Auth::check() && $post->isAuthor(Auth::user()))
 			<a href="{{{ action('PostsController@edit', $post->id) }}}"><span class="badge">Edit post</span></a>
 			{{ Form::open(array('action' => array('PostsController@destroy', $post->id), 'method' => 'DELETE')) }}
 				{{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
